Reorganized code into packages for readability
